Added support for TGCN (#9)
This commit focuses on adding TGCN support to Seastar. TGCN model definition and training code has been added in the exp/tgcn folder. Most of the commit is attributed to adding gradient accumulation capabilities which is detrimental to temporal models but is missing in seastar. The changes required were

Add a stack data structure in executor to keep track of the state of a particular execution of the seastar cell. This is essential since during backdrop of that cell the state is necessary.

Use the tensor_map to track the state of current execution of seastar cell and push it at the end of forward propagation. This is then popped at the backward propagation of the corresponding seastar cell.

Manage stack operations such that there isn't a build up of memory.

The model compiles and trains with capabilities parallel to DGL-TGCN. It can be noted that each epoch takes 4.5s in DGL-TGN but it takes only 2s in Seastar-TGN. Additionally we have verified the correctness by comparing the arg values during forward and backward propagation of a random execution time in between an epoch.