:tada: Initial commit
Added the CUDA file gpma.cu that contains the code for the GPMA data structure. It also contains PyBind11 python bindings for the GPMA class. As of now, the bindings work properly.

Added a README that has the command to build the gpma.cu file into a shared library file gpma.so