➕ Introduced Boorah dataset for Temporal GNNs
Instead of WikiMath we will be working on Boorah datasets, which is similar to the WikiMath graph dataset. Created a script to generate the dataset in accordance to the number of nodes and edge multiplier passed into the program by the user.

We have run Seastar and PyG-T on Boorah and as of now our Seastar is 2 times or 100% faster than PyG-T

Will be testing it on multiple sized Boorah datasets.