🎉 Initial setup to benchmark GPMA
In the process of speeding up GPMA's copy_label_edges method, a new folder called seastar-gpma has been created inside the benchmarking directory.

Complete instructions on how to run the benchmark has been provided inside the README.md. As of now it outputs the total time taken to run and the time taken to copy the label edges for each dataset.

Our goal is to speed up the copy_label_edges functions speed.