:wheelchair: Simplified backend framework specification in seastar function decorator
In order to streamline the usage of the @seastar function decorator, we have made an improvement to the backend framework specification. Previously, it was necessary to mention the backend framework in the decorator itself. However, we have eliminated this requirement, simplifying the usage for users.

Now, users only need to specify the backend framework they will be using once while creating the GNN layer. No additional information needs to be passed elsewhere. This change reduces complexity and enhances the user experience.