✅ Seastar Backend, Val and Op abstraction complete
Added abstraction using Factory Class design pattern for the backend wrappers, Val and Op. Made import changes wherever required. Structured the code in a well organized manner.

As of now the only backend framework that is supported is PyTorch. But the code is organized in a manner that adding support for other frameworks such as tensorflow and MXNet would be straightforward. This can be left as a future implementation.