:bug: Resolved update_gpma OOM bug
It seems that PyBind11 is creating some issue when cudaMalloc is called from device code. This is issue is encountered with rebalancing_kernel device function. The fix was to allocate the memory in the host code and then call the device function, which updates its pointer based on the blockIdx. Was successful in executing the code for WikiTalkTemporal dataset.