➕Degree sorting added to GPMAGraph
Degree sorting improved GPMA performance from 0.37 to 0.31 for feature size 1 on mathoverflow dataset with slice 10.0. The time taken for sorting per snapshot is between 0.0005-0.001. This is a sufficiently negligible range and as we can see we do see benifits in performance by adding this support.

Additionally, Seastar-Plus now fully support degree sorting as well. Static, Naive, PCSR and GPMA have implementations for this feature. The performance and correctness of this implementation have been verified as well.