📝 Added references to dataset docstrings
Added reference links in the docstrings which shows up in the documentation. So far it's the links to the dataset loader provided by PyG-T from where we got the datasets.

Also removed all the private attributes from the attributes list in the docstrings. Added gdata to the attributes list.

Provided the return type for the getter methods in the dataset loaders