🎉 Made modifications to the dataset tests
Removed the test which contains the url argument, since we removed the url argument from the dataloader. And added tests which contains the redownload argument.

Added argument type and value checking in EnglandCovifDataLoader. Added the necessary exception error tests for the same.

Made changes in the train.py and utils.py to reflect the new changes in the dataloader API. Rest of the training code remains the same.